requirements.txt
# PyTorch with CUDA 11.8 support
torch==2.0.1+cu118
torchvision==0.15.2+cu118
torchaudio==2.0.2+cu118

# Transformers (Hugging Face)
transformers==4.31.0
tokenizers==0.13.3

# Streamlit for frontend
streamlit==1.31.1

# Optional: For PIL image processing
Pillow

# Optional: Streamlit enhancements
streamlit-option-menu
streamlit-extras
